@@ -6,6 +6,7 @@ using System.Threading.Tasks;
 using Microsoft.AspNetCore;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 
 namespace MyCourse
@@ -17,8 +18,16 @@ namespace MyCourse
             CreateWebHostBuilder(args).Build().Run();
         }
 
-        public static IWebHostBuilder CreateWebHostBuilder(string[] args) =>
-            WebHost.CreateDefaultBuilder(args)
-                .UseStartup<Startup>();
+        //public static IWebHostBuilder CreateWebHostBuilder(string[] args) =>
+        public static IHostBuilder CreateWebHostBuilder(string[] args) =>
+            //WebHost.CreateDefaultBuilder(args)
+            //.UseStartup<Startup>();
+            //webHost sostituito dalla versione .Net 3.0 da un Generic Host
+            Host.CreateDefaultBuilder(args)
+            .ConfigureWebHostDefaults(webHostBuilder => {
+                webHostBuilder.UseStartup<Startup>();
+            });
+            //se volessi configurare la dependency injection in un'applicazione console userei: 
+            //.ConfigureServices            
     }
 }
